# openapi.yaml
openapi: 3.1.0
info:
  title: Astradio API
  version: "1.0"
servers:
  - url: https://api.astradio.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, username]
      properties:
        id: { type: string, format: ulid }
        email: { type: string, format: email }
        username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string }
    Profile:
      type: object
      properties:
        birth_datetime: { type: string, format: date-time }
        birth_lat: { type: number }
        birth_lon: { type: number }
        privacy_level: { type: string, enum: [public, friends, private] }
        compare_opt_in: { type: boolean }
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      summary: Get current user
      responses:
        "200": { description: Me, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
    patch:
      summary: Update account fields
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/User" }
      responses: { "200": { description: Updated } }
  /users/me/profile:
    patch:
      summary: Update profile fields
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Profile" }
      responses: { "200": { description: Updated } }
  /users/{username}:
    get:
      summary: Public profile
      parameters: [{ name: username, in: path, required: true, schema: {type: string} }]
      responses: { "200": { description: OK } }
...